+++ docs/src/changelog.md
@@ -21,6 +21,7 @@ If your currently installed *Quality-time* version is not the penultimate versio
 - Clicking the target value expands the metric on the metric configuration tab, and collapses the metric when clicked again while the configuration tab is active. Closes [#13034](https://github.com/ICTU/quality-time/issues/13034).
 - Clicking the unit expands the metric on the metric configuration tab, and collapses the metric when clicked again while the configuration tab is active. Closes [#13036](https://github.com/ICTU/quality-time/issues/13036).
 - Clicking the time left expands the metric on the technical debt tab, and collapses the metric when clicked again while the technical debt tab is active. Closes [#13038](https://github.com/ICTU/quality-time/issues/13038).
+- Clicking the metric comment expands the metric on the technical debt tab, and collapses the metric when clicked again while the technical debt tab is active. Closes [#13040](https://github.com/ICTU/quality-time/issues/13040).
 
 ## v5.53.0 - 2026-04-24